Add deposit withdrawal logic

+++ src/services/DepositService.js
@@ -1,6 +1,9 @@
 const Deposit = require('../models/deposit')
 
 class DepositService {
+  REVOCABLE_TYPE = 1
+  IRREVOCABLE_TYPE = 2
+
   async getAllDeposits () {
     const deposits = await Deposit.find({})
       .populate('owner')
@@ -9,22 +12,40 @@ class DepositService {
     return deposits
   }
 
-  async getDepositById (_id) {
+  async withdrawDeposit (_id) {
     const deposit = await Deposit.findOne({ _id })
+    let withdrawAmount
+    if (deposit.type === this.REVOCABLE_TYPE) {
+      withdrawAmount = deposit.balance - deposit.sum
+      deposit.balance = deposit.sum
+      await deposit.save()
 
-    return deposit
+      return {
+        message: `Withdrew ${withdrawAmount}!`,
+        style: 'success'
+      }
+    } else if (deposit.type === this.IRREVOCABLE_TYPE) {
+      withdrawAmount = deposit.balance
+
+      await Deposit.deleteOne({ _id })
+
+      return {
+        message: `Withdrew ${withdrawAmount} and closed deposit!`,
+        style: 'success'
+      }
+    } else {
+      return {
+        message: 'Incorrect deposit type!',
+        type: 'danger'
+      }
+    }
   }
 
   async createDeposit (data) {
-    delete data._method
-
     return Deposit.create({
-      ...data
+      ...data,
+      balance: data.sum
     })
-  }
-
-  async deleteDepositById (_id) {
-    await Deposit.deleteOne({ _id })
   }
 }
